Use specific times for evaluating journeys
To ensure consistency, using hte smae time each week for journeys. This
is noon on the next monday, a pretty neautral and uneventful time.

Could add options to set time on front end, vut that is extra choices
for the user.

Also added avility to use a range of times/days and selectd the best. This is for public transport only. It is required because available
public transport can vary so much, and can be affected by events such as
strikes...

Not done driving as less affected and so saves requesr volume, but could
enable it in future.